% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoEncoders.R
\name{drugAuto}
\alias{drugAuto}
\title{Implement the drug autoencoder}
\usage{
drugAuto(
  trainData,
  featureEngineeringSettings = list(drugAnalysisId = 301, drugCodeLoc =
    system.file("models/drugs_codes_ALL_OF_US.pkl", package =
    "GlaucomaPrescreeningPrediction"), autoEncoderLoc =
    system.file("models/drugs_autoencoder_model_ALL_OF_US.h5", package =
    "GlaucomaPrescreeningPrediction"))
)
}
\arguments{
\item{trainData}{The training data to apply the autoencoder to}

\item{featureEngineeringSettings}{settings for loading the autoencoder}
}
\value{
The plp data with the auto encoder features added for the drugs
}
\description{
Call the drug autoencoder to get 128 features
}
\details{
Used by applyFeatureEngineering to add the drug autoencoders
}
