% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execute}
\alias{execute}
\title{Validates the glaucoma model}
\usage{
execute(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  tempEmulationSchema = Sys.getenv("DATABRICKS_SCRATCH_SCHEMA"),
  cohortTable = "glau_screen_cohort",
  targetId = 23884,
  outcomeId = 23933,
  sampleSize = NULL
)
}
\arguments{
\item{connectionDetails}{The connection details for the OMOP CDM data}

\item{cdmDatabaseSchema}{The schema containing the OMOP CDM data}

\item{cohortDatabaseSchema}{A schema with read/write access that contains the target and outcome cohorts}

\item{tempEmulationSchema}{The schema to use for temp tables if the dbms does not support temp tables}

\item{cohortTable}{The name of the cohort table that contains the target and outcome cohorts}

\item{targetId}{The validation target cohort definition id}

\item{outcomeId}{The validation outcome cohort definition id}

\item{sampleSize}{Number of patients to sample from the full target cohort}
}
\value{
A list with the prediction data.frame and evaluation list and the data
}
\description{
Validates the glaucoma model
}
\details{
The user specifies the connectionDetails to the new OMOP CDM data,
}
