% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{executeGlaucoma}
\alias{executeGlaucoma}
\title{Executes all the analyses}
\usage{
executeGlaucoma(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  tempEmulationSchema = Sys.getenv("DATABRICKS_SCRATCH_SCHEMA"),
  cohortTable = "glau_screen_cohort",
  outputFolder = tempdir(),
  minCellCount = 0,
  runCohortGeneration = TRUE,
  runCharacterization = TRUE,
  runValidation = TRUE,
  runFineTune = TRUE,
  runModelDevelopment = TRUE,
  sampleSize = NULL
)
}
\arguments{
\item{connectionDetails}{The connection details for the OMOP CDM data}

\item{cdmDatabaseSchema}{The schema containing the OMOP CDM data}

\item{cohortDatabaseSchema}{A schema with read/write access that contains the target and outcome cohorts}

\item{tempEmulationSchema}{The schema to use for temp tables if the dbms does not support temp tables}

\item{cohortTable}{The name of the cohort table that contains the target and outcome cohorts}

\item{outputFolder}{The location of the folder to save all the results to}

\item{minCellCount}{The minimum cell count to save (otherwise it is saved as -minCellCount to represent <)}

\item{runCohortGeneration}{Whether to run the cohort generation}

\item{runCharacterization}{Whether to run the characterization}

\item{runValidation}{Whether to run the model validation}

\item{runFineTune}{Whether to run the model fineTuning}

\item{runModelDevelopment}{Whether to run a benchmark LR model}

\item{sampleSize}{Number of patients to sample from the full target cohort for validation and development}
}
\value{
A location where all the results are saved to.
}
\description{
Executes all the analyses
}
\details{
The user specifies the connectionDetails to the new OMOP CDM data,
}
