% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateCohorts.R
\name{generateCohorts}
\alias{generateCohorts}
\title{Generate study cohorts}
\usage{
generateCohorts(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTableName = "glaucoma_prescreen",
  tempEmulationSchema = NULL
)
}
\arguments{
\item{connectionDetails}{The OMOP CDM connection details created using DatabaseConnector::createConnectionDetails}

\item{cdmDatabaseSchema}{The schema containing the OMOP CDM}

\item{cohortDatabaseSchema}{The schema to add the cohort table to}

\item{cohortTableName}{The name of the cohort table to create for this study}

\item{tempEmulationSchema}{Optional temp schema}
}
\value{
TRUE
}
\description{
Generate study cohorts
}
\details{
The user specifies the connection details for the OMOP CDM, the CDM database schema
the cohort table schema and the name of the table you want the cohorts to be generated into
optionally you can also enter the temp schema that is used to create temporary tables.
}
