% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plpModel.R
\name{getModel}
\alias{getModel}
\title{Extracts measurement covariates}
\usage{
getModel(
  modelName = "model_ALL_OF_US.h5",
  conditionFile = "diag_codes_ALL_OF_US.pkl",
  conditionAutoFile = "diag_autoencoder_model_ALL_OF_US.h5",
  drugFile = "drugs_codes_ALL_OF_US.pkl",
  drugAutoFile = "drugs_autoencoder_model_ALL_OF_US.h5"
)
}
\arguments{
\item{modelName}{filename of the model}

\item{conditionFile}{filename of the pickle with the condition codes}

\item{conditionAutoFile}{filename of the condition autoencoder}

\item{drugFile}{filename of the pickle with the drug codes}

\item{drugAutoFile}{filename of the drug autoencoder}
}
\value{
CovariateData object with covariates, covariateRef, and analysisRef tables
}
\description{
Extracts measurement covariates
}
\details{
The user specifies a cohort and time period and then a covariate is constructed whether they are in the
cohort during the time periods relative to target population cohort index
}
