% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurementFeatures.R
\name{implementAgeScale}
\alias{implementAgeScale}
\title{function to scale age 1002 - min:21	max:109	range:88}
\usage{
implementAgeScale(trainData, featureEngineeringSettings, model = NULL)
}
\arguments{
\item{trainData}{The training data to apply the autoencoder to}

\item{featureEngineeringSettings}{settings for loading the autoencoder}

\item{model}{The plp model}
}
\value{
The plp data with the scaled age added as covariate id 2002
}
\description{
Call the age scaling function
}
\details{
Used by applyFeatureEngineering to scale the age in years
}
