% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Characterization.R
\name{runCharacterization}
\alias{runCharacterization}
\title{run characterization}
\usage{
runCharacterization(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  tempEmulationSchema = Sys.getenv("DATABRICKS_SCRATCH_SCHEMA"),
  cohortTable = "glau_screen_cohort",
  minCellCount = 0,
  saveLoc = tempdir()
)
}
\arguments{
\item{connectionDetails}{The connection details for the OMOP CDM data}

\item{cdmDatabaseSchema}{The schema containing the OMOP CDM data}

\item{cohortDatabaseSchema}{A schema with read/write access that contains the target and outcome cohorts}

\item{tempEmulationSchema}{The schema to use for temp tables if the dbms does not support temp tables}

\item{cohortTable}{The name of the cohort table that contains the target and outcome cohorts}

\item{minCellCount}{The minCellCount for results}

\item{saveLoc}{The location to save the results to}
}
\value{
The location where the csv files for the characterization are
}
\description{
run characterization
}
\details{
The user specifies the connectionDetails to the new OMOP CDM data,
}
