% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{validationModel}
\alias{validationModel}
\title{Validates the glaucoma model}
\usage{
validationModel(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  tempEmulationSchema = Sys.getenv("DATABRICKS_SCRATCH_SCHEMA"),
  cohortTable = "glau_screen_cohort",
  targetId = 23884,
  outcomeId = 23933,
  sampleSize = NULL,
  model = c("ALL_OF_US_8_17_26_221", "ALL_OF_US")[1],
  modelType = c(".keras", ".h5")[1]
)
}
\arguments{
\item{connectionDetails}{The connection details for the OMOP CDM data}

\item{cdmDatabaseSchema}{The schema containing the OMOP CDM data}

\item{cohortDatabaseSchema}{A schema with read/write access that contains the target and outcome cohorts}

\item{tempEmulationSchema}{The schema to use for temp tables if the dbms does not support temp tables}

\item{cohortTable}{The name of the cohort table that contains the target and outcome cohorts}

\item{targetId}{The validation target cohort definition id}

\item{outcomeId}{The validation outcome cohort definition id}

\item{sampleSize}{Number of patients to sample from the full target cohort}

\item{model}{The name of the model to run}

\item{modelType}{either '.keras' or '.h5'}
}
\value{
A list with the prediction data.frame and evaluation list and the data
}
\description{
Validates the glaucoma model
}
\details{
The user specifies the connectionDetails to the new OMOP CDM data,
}
