% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoEncoders.R
\name{condAuto}
\alias{condAuto}
\title{Implement the condition autoencoder}
\usage{
condAuto(
  trainData,
  featureEngineeringSettings = list(conditionAnalysisId = 101, conditionCodeLoc =
    system.file("models/diag_codes.pkl", package = "GlaucomaPrescreeningPrediction"),
    autoEncoderLoc = system.file("models/diag_autoencoder_model_ALL_OF_US.h5", package =
    "GlaucomaPrescreeningPrediction"))
)
}
\arguments{
\item{trainData}{The training data to apply the autoencoder to}

\item{featureEngineeringSettings}{settings for loading the autoencoder}
}
\value{
The plp data with the auto encoder features added
}
\description{
Call the condition autoencoder to get 128 features
}
\details{
Used by applyFeatureEngineering to add the condition autoencoders
}
