% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlaucomaPrescreeningPrediction.R
\docType{package}
\name{GlaucomaPrescreeningPrediction}
\alias{GlaucomaPrescreeningPrediction-package}
\alias{GlaucomaPrescreeningPrediction}
\title{GlaucomaPrescreeningPrediction}
\description{
A package for implementing an existing glaucoma screening model.
}
\author{
\strong{Maintainer}: Jenna Reps \email{jreps@its.jnj.com}

}
